refactor(settings): house sidebar components under components/Settings
Move the settings page's sub-components — SettingsNavigation,
SettingsSearchInput, ProjectSettings, CreateProjectSettings — from
pages/ into components/Settings/, leaving Settings.tsx as the only
settings entry in pages/. Pare the sidebar search back to the
essentials: drop the "/" focus hotkey and Escape handling, keeping the
fuzzy filter, keyword aliases, match highlighting and empty state.
