+++ packages/frontend/src/Routes.tsx
@@ -7,6 +7,7 @@ import JobDetailsDrawer from './components/JobDetailsDrawer';
 import NavBar from './components/NavBar';
 import PrivateRoute from './components/PrivateRoute';
 import ProjectRoute from './components/ProjectRoute';
+import CreateProjectSettings from './components/Settings/CreateProjectSettings';
 import UserCompletionModal from './components/UserCompletionModal';
 import FunnelBuilder from './features/funnelBuilder/FunnelBuilderPage';
 import { MetricCatalogView } from './features/metricsCatalog/types';
@@ -15,7 +16,6 @@ import AppPreviewTest from './pages/AppPreviewTest';
 import AuthPopupResult from './pages/AuthPopupResult';
 import ChartHistory from './pages/ChartHistory';
 import CreateProject from './pages/CreateProject';
-import CreateProjectSettings from './pages/CreateProjectSettings';
 import Dashboard from './pages/Dashboard';
 import DashboardHistory from './pages/DashboardHistory';
 import Explorer from './pages/Explorer';
