+++ packages/frontend/src/components/Settings/CreateProjectSettings.tsx
@@ -2,10 +2,10 @@ import { Stack, Text, Title } from '@mantine/core';
 import { useQueryClient } from '@tanstack/react-query';
 import { type FC } from 'react';
 import { useNavigate, useParams } from 'react-router';
-import Page from '../components/common/Page/Page';
-import PageSpinner from '../components/PageSpinner';
-import ProjectTablesConfiguration from '../components/ProjectTablesConfiguration/ProjectTablesConfiguration';
-import useApp from '../providers/App/useApp';
+import useApp from '../../providers/App/useApp';
+import Page from '../common/Page/Page';
+import PageSpinner from '../PageSpinner';
+import ProjectTablesConfiguration from '../ProjectTablesConfiguration/ProjectTablesConfiguration';
 
 const CreateProjectSettings: FC = () => {
     const { projectUuid } = useParams<{ projectUuid: string }>();
