+++ packages/frontend/src/components/Settings/ProjectSettings.tsx
@@ -1,33 +1,33 @@
 import { Stack } from '@mantine-8/core';
 import { useMemo, type FC } from 'react';
 import { Navigate, useParams, useRoutes, type RouteObject } from 'react-router';
-import { DocumentTitle } from '../components/common/DocumentTitle';
-import ErrorState from '../components/common/ErrorState';
-import PageBreadcrumbs from '../components/common/PageBreadcrumbs';
-import SuboptimalState from '../components/common/SuboptimalState/SuboptimalState';
-import CompilationHistory from '../components/CompilationHistory';
-import { DataOps } from '../components/DataOps';
-import { DefaultUserSpaces } from '../components/DefaultUserSpaces';
-import { useIsGitProject } from '../components/Explorer/WriteBackModal/hooks';
-import PreAggregateAudit from '../components/PreAggregateAudit';
-import PreAggregateMaterializations from '../components/PreAggregateMaterializations';
-import ProjectUserAccess from '../components/ProjectAccess';
-import ProjectAppearance from '../components/ProjectAppearance/ProjectAppearance';
-import { UpdateProjectConnection } from '../components/ProjectConnection';
-import ProjectParameters from '../components/ProjectParameters';
-import ProjectPreviewExpiration from '../components/ProjectPreviewExpiration';
-import ProjectTablesConfiguration from '../components/ProjectTablesConfiguration/ProjectTablesConfiguration';
-import SettingsQueryTimezone from '../components/SettingsQueryTimezone';
-import SettingsScheduler from '../components/SettingsScheduler';
-import SettingsUsageAnalytics from '../components/SettingsUsageAnalytics';
-import { SettingsValidator } from '../components/SettingsValidator';
-import VerifiedContentPanel from '../components/VerifiedContent/VerifiedContentPanel';
-import SettingsEmbed from '../ee/features/embed/SettingsEmbed';
-import { ProjectChangesets } from '../features/changesets/components/ProjectChangesets';
-import PullRequestsPage from '../features/pullRequests/components/PullRequestsPage';
-import RecentlyDeletedPage from '../features/recentlyDeleted/components/RecentlyDeletedPage';
-import { useProject } from '../hooks/useProject';
-import useApp from '../providers/App/useApp';
+import SettingsEmbed from '../../ee/features/embed/SettingsEmbed';
+import { ProjectChangesets } from '../../features/changesets/components/ProjectChangesets';
+import PullRequestsPage from '../../features/pullRequests/components/PullRequestsPage';
+import RecentlyDeletedPage from '../../features/recentlyDeleted/components/RecentlyDeletedPage';
+import { useProject } from '../../hooks/useProject';
+import useApp from '../../providers/App/useApp';
+import { DocumentTitle } from '../common/DocumentTitle';
+import ErrorState from '../common/ErrorState';
+import PageBreadcrumbs from '../common/PageBreadcrumbs';
+import SuboptimalState from '../common/SuboptimalState/SuboptimalState';
+import CompilationHistory from '../CompilationHistory';
+import { DataOps } from '../DataOps';
+import { DefaultUserSpaces } from '../DefaultUserSpaces';
+import { useIsGitProject } from '../Explorer/WriteBackModal/hooks';
+import PreAggregateAudit from '../PreAggregateAudit';
+import PreAggregateMaterializations from '../PreAggregateMaterializations';
+import ProjectUserAccess from '../ProjectAccess';
+import ProjectAppearance from '../ProjectAppearance/ProjectAppearance';
+import { UpdateProjectConnection } from '../ProjectConnection';
+import ProjectParameters from '../ProjectParameters';
+import ProjectPreviewExpiration from '../ProjectPreviewExpiration';
+import ProjectTablesConfiguration from '../ProjectTablesConfiguration/ProjectTablesConfiguration';
+import SettingsQueryTimezone from '../SettingsQueryTimezone';
+import SettingsScheduler from '../SettingsScheduler';
+import SettingsUsageAnalytics from '../SettingsUsageAnalytics';
+import { SettingsValidator } from '../SettingsValidator';
+import VerifiedContentPanel from '../VerifiedContent/VerifiedContentPanel';
 
 const ProjectSettings: FC = () => {
     const { projectUuid } = useParams<{
