+++ packages/frontend/src/components/Settings/SettingsSearchInput.tsx
@@ -1,34 +1,27 @@
-import { ActionIcon, Kbd, TextInput } from '@mantine-8/core';
+import { ActionIcon, TextInput } from '@mantine-8/core';
 import { IconSearch, IconX } from '@tabler/icons-react';
-import { forwardRef } from 'react';
-import MantineIcon from '../components/common/MantineIcon';
+import { type FC } from 'react';
+import MantineIcon from '../common/MantineIcon';
 
 type SettingsSearchInputProps = {
     value: string;
     onChange: (value: string) => void;
 };
 
-const SettingsSearchInput = forwardRef<
-    HTMLInputElement,
-    SettingsSearchInputProps
->(({ value, onChange }, ref) => {
+const SettingsSearchInput: FC<SettingsSearchInputProps> = ({
+    value,
+    onChange,
+}) => {
     const hasValue = value.length > 0;
 
     return (
         <TextInput
-            ref={ref}
             size="xs"
             radius="md"
             placeholder="Search settings"
             aria-label="Search settings"
             value={value}
             onChange={(event) => onChange(event.currentTarget.value)}
-            onKeyDown={(event) => {
-                if (event.key === 'Escape' && hasValue) {
-                    event.preventDefault();
-                    onChange('');
-                }
-            }}
             leftSection={<MantineIcon icon={IconSearch} color="ldGray.6" />}
             rightSectionPointerEvents={hasValue ? 'auto' : 'none'}
             rightSection={
@@ -42,14 +35,10 @@ const SettingsSearchInput = forwardRef<
                     >
                         <MantineIcon icon={IconX} />
                     </ActionIcon>
-                ) : (
-                    <Kbd size="xs">/</Kbd>
-                )
+                ) : null
             }
         />
     );
-});
-
-SettingsSearchInput.displayName = 'SettingsSearchInput';
+};
 
 export default SettingsSearchInput;
