+++ packages/frontend/src/hooks/settings/CLAUDE.md
@@ -41,21 +41,30 @@ if (isMyFeatureEnabled && ability?.can('manage', 'Organization')) {
 <importantToKnow>
 - **The sidebar and the router are NOT unified yet.** A nav entry only adds the sidebar link. To make the page reachable you must ALSO add a matching route:
   - Org/user-level pages → the `routes` `useMemo` in `@/packages/frontend/src/pages/Settings.tsx`.
-  - `current-project` pages (`to` under `/projectManagement/${projectUuid}/...`) → routes live in `@/packages/frontend/src/pages/ProjectSettings.tsx`, not in `Settings.tsx`.
+  - `current-project` pages (`to` under `/projectManagement/${projectUuid}/...`) → routes live in `@/packages/frontend/src/components/Settings/ProjectSettings.tsx`, not in `Settings.tsx`.
   Keep the entry's gating in sync with the route's gating, or you get a nav link to a forbidden/missing route, or a reachable route with no nav link. (Unifying both into one registry is the remaining part of SPK-469.)
 - `useSettingsContext` is the shared primitive: `Settings.tsx` reads it for its `routes` memo and render guards, and `useSettingsNavigation` reads it for the nav. If a route needs a flag the context doesn't expose yet, add it to `SettingsContext` and return it from `useSettingsContext`.
 - The nav model is built in a `useMemo` that runs before the page's `!user` guard, so it uses `user?.ability` (optional). It's only consumed after the guard, so transient empty sections are harmless — keep the optional chaining.
 - Changing a `label` here is purely cosmetic; changing a `to` requires updating the corresponding route path too.
 </importantToKnow>
 
+<sidebarInteraction>
+The sidebar's search is data-driven (no DOM walking) and its state lives directly in `Settings.tsx`:
+
+- `search` (raw, feeds the input) plus a debounced copy (feeds `filterSettingsNavigation`). The search resets on navigation, computed during render so it never lags the route change.
+- `isSidebarCollapsed` — a plain `useState` toggling the collapsible panel; not persisted.
+
+`SettingsNavigation` takes `sections` + `searchQuery`. Group expansion stays uncontrolled via Mantine's `defaultOpened` (open on the active route, or forced open while filtering); the nav remounts on the filtering toggle so `defaultOpened` re-evaluates. Matching label text is highlighted with Mantine's `<Highlight>`.
+</sidebarInteraction>
+
 <links>
 - Gating context: @/packages/frontend/src/hooks/settings/useSettingsContext.ts
 - Nav model assembly: @/packages/frontend/src/hooks/settings/useSettingsNavigation.ts
 - Search filter (Fuse.js): @/packages/frontend/src/hooks/settings/filterSettingsNavigation.ts
 - Entry/section + context types: @/packages/frontend/src/hooks/settings/types.ts
-- Renderer: @/packages/frontend/src/pages/SettingsNavigation.tsx
-- Search input: @/packages/frontend/src/pages/SettingsSearchInput.tsx
+- Renderer: @/packages/frontend/src/components/Settings/SettingsNavigation.tsx
+- Search input: @/packages/frontend/src/components/Settings/SettingsSearchInput.tsx
 - Router + org/user routes: @/packages/frontend/src/pages/Settings.tsx
-- Project-level routes: @/packages/frontend/src/pages/ProjectSettings.tsx
+- Project-level routes: @/packages/frontend/src/components/Settings/ProjectSettings.tsx
 - Permissions/abilities: @/packages/common/src/authorization/
 </links>
